#9105: move state controller to content script
