+++ src/starterBricks/sidebar/sidebarStarterBrick.test.ts
@@ -29,7 +29,6 @@ import {
   isSidePanelOpen,
   sidebarShowEvents,
 } from "@/contentScript/sidebarController";
-import { setState } from "@/platform/state/stateController";
 import {
   modComponentRefFactory,
   modMetadataFactory,
@@ -188,7 +187,7 @@ describe("sidebarExtension", () => {
 
     expect(rootReader.readCount).toBe(0);
 
-    setState({
+    getPlatform().state.setState({
       namespace: StateNamespaces.MOD,
       data: {},
       mergeStrategy: MergeStrategies.REPLACE,
@@ -210,7 +209,7 @@ describe("sidebarExtension", () => {
     // Runs because statechange mods also run on manual
     expect(rootReader.readCount).toBe(1);
 
-    setState({
+    getPlatform().state.setState({
       namespace: StateNamespaces.MOD,
       // Data needs to be different than previous to trigger a state change event
       data: { foo: 42 },
@@ -226,7 +225,7 @@ describe("sidebarExtension", () => {
     expect(rootReader.readCount).toBe(2);
 
     // Should ignore state change from other mod
-    setState({
+    getPlatform().state.setState({
       namespace: StateNamespaces.MOD,
       data: {},
       mergeStrategy: MergeStrategies.REPLACE,
@@ -274,7 +273,7 @@ describe("sidebarExtension", () => {
     expect(rootReader.readCount).toBe(1);
 
     for (let i = 0; i < 10; i++) {
-      setState({
+      getPlatform().state.setState({
         namespace: StateNamespaces.MOD,
         data: { foo: i },
         mergeStrategy: MergeStrategies.REPLACE,
